requirements.txt
google-cloud-bigquery==2.30.0
